server/requirements.txt
# A place to list your Python dependencies.
#
# As a reminder, to create a virtual environment:
# python -m venv .venv
# source .venv/bin/activate
#
# Install dependencies:
# pip install -r requirements.txt

fastapi[standard]==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
black==23.11.0
requests==2.31.0
debugpy==1.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
google-generativeai==0.3.2
python-dotenv==1.0.0
httpx==0.25.2
